chore: refactor replaceHistoryLocation
